Beta 4.3 Bugfix
Fixed Name Error from CaultronBrewingProcedure to CauldronBrewingProcedure.

Fixed Version error via update (was 3-1 when it was supposed to be 3.1, that is fixed in future versions)

Fixed bottle duplication and no item loss in cauldron brewing.

Fixed Xray glitch with cauldrons.

Added more biomes for the Paeonia Flower to spawn in since it was too rare. Now spawning in taigas, tall birch forests, and swamp hills (along with more biomes)

Next Beta is a Pre-Release for 0.7.0